minor changes for debugging

+++ srcs/Server/server.hpp
@@ -36,9 +36,9 @@ private:
 	std::map<std::string, setter>	_typeFunctionMap;
 	std::vector<location>			_locations;
 
-	int					_socketFd;
+	long				_socketFd;
 	struct sockaddr_in	_addr;
-	int					_acceptFd;
+	long				_acceptFd;
 
 public:
 	server();
@@ -67,9 +67,9 @@ public:
 	const std::vector<std::string>	&getIndices() const;
 	const std::vector<location>		&getLocations() const;
 
-	const int						&getSocketFd() const;
+	const long						&getSocketFd() const;
 	const struct sockaddr_in		&getAddr() const;
-	const int						&getAcceptFd() const;
+	const long						&getAcceptFd() const;
 
 	void		addLocation(location &newLoc);
 	bool		valueCheck() const;
